backend/pyproject.toml
[project]
name = "music-graph-backend"
version = "0.1.0"
description = "FastAPI backend for Yandex Music graph visualization"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.3",
  "arq>=0.26.1",
  "asyncpg>=0.30.0",
  "cryptography>=43.0.3",
  "fastapi>=0.115.0",
  "greenlet>=3.1.1",
  "httpx>=0.27.2",
  "psycopg[binary]>=3.2.3",
  "pydantic>=2.9.2",
  "pydantic-settings>=2.5.2",
  "PyJWT>=2.9.0",
  "redis>=5.1.1",
  "SQLAlchemy>=2.0.35",
  "uvicorn[standard]>=0.31.1",
  "ya-passport-auth>=0.1.0",
  "yandex-music>=3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.7.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
